+++ sdk/resourcemanager/compute/armbulkactions/models_serde.go
@@ -2104,8 +2104,8 @@ func (i *InnerError) UnmarshalJSON(data []byte) error {
 	return nil
 }
 
-// MarshalJSON implements the json.Marshaller interface for type InnerError.
-func (i InnerError) MarshalJSON() ([]byte, error) {
+// MarshalJSON implements the json.Marshaller interface for type BulkInstancesInnerError.
+func (i BulkInstancesInnerError) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)
 	populate(objectMap, "code", i.Code)
 	populate(objectMap, "errorDetail", i.ErrorDetail)
@@ -2114,8 +2114,8 @@ func (i InnerError) MarshalJSON() ([]byte, error) {
 	return json.Marshal(objectMap)
 }
 
-// UnmarshalJSON implements the json.Unmarshaller interface for type InnerError.
-func (i *InnerError) UnmarshalJSON(data []byte) error {
+// UnmarshalJSON implements the json.Unmarshaller interface for type BulkInstancesInnerError.
+func (i *BulkInstancesInnerError) UnmarshalJSON(data []byte) error {
 	var rawMsg map[string]json.RawMessage
 	if err := json.Unmarshal(data, &rawMsg); err != nil {
 		return fmt.Errorf("unmarshalling type %T: %s", i, err.Error())